// src/types.ts
/**
 * Options for testing PrimeNG button component
 */
export interface NgButtonOptions {
  /**
   * Whether the button is disabled
   */
  disabled?: boolean;
  
  /**
   * Whether to click the button
   */
  click?: boolean;
  
  /**
   * Expected label text of the button
   */
  expectLabel?: string;
}